refator: Version.clientVersion to cli-common as VersionHelper
