+++ bom/application/pom.xml
@@ -6672,6 +6672,11 @@
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
+                <artifactId>quarkus-cli-common</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.quarkus</groupId>
                 <artifactId>quarkus-analytics-common</artifactId>
                 <version>${project.version}</version>
             </dependency>
